Minor cleanup - define out some code which isn't used currently (probably could be kicked out) - just some minor type fixes Just stuff I find while hunting while valgrind is still giving me errors (even thought less now, generally when ai makes second round or after loading a few times).

+++ src/game.cpp
@@ -1078,7 +1078,7 @@ void Game::Update()
     guiHud.SetTimeTrack(time-mRaceStartTime);
 
     PROFILE_START(200);
-    for ( unsigned int p=0; p<mPlayers.size(); ++p)
+    for ( size_t p=0; p<mPlayers.size(); ++p)
     {
         mPlayers[p]->PrePhysicsUpdate(time);
     }
